+++ src/pages/Explore/ui/CategorySidebar/index.tsx
@@ -10,7 +10,7 @@ import { ExpandDownSVG } from '@/shared/assets/SVGicons';
 import { kebabToTitleCase } from '@/shared/lib';
 import { QUERY_KEYS } from '@/shared/query/key';
 import { useFetchQuery } from '@/shared/query/useFetchQuery';
-import { CategoryListResponse } from '@/types/response';
+import { CategoryListResponse } from '@/shared/types/response';
 
 const CategorySidebar = () => {
   const productId = useSelector((state: RootState) => state.productId);
