+++ src/pages/Mypage/ui/AddressCard.tsx
@@ -1,5 +1,5 @@
 import * as S from '@/pages/Mypage/MyPage.styled';
-import { AddressType } from '@/types/response/user.types';
+import { AddressType } from '@/shared/types/response';
 
 const AddressCard = ({
   address,
