+++ src/pages/Mypage/ui/Cart/index.tsx
@@ -4,7 +4,7 @@ import * as S from '@/pages/Mypage/MyPage.styled';
 import { SERVICE_URLS } from '@/shared/api/endpoints';
 import { QUERY_KEYS } from '@/shared/query/key';
 import { useFetchQuery } from '@/shared/query/useFetchQuery';
-import { CartResponse } from '@/types/response/cart.types';
+import { CartResponse } from '@/shared/types/response';
 
 const ItemRow = styled.div`
   display: grid;
