+++ src/pages/Mypage/ui/FinanceCard.tsx
@@ -1,5 +1,5 @@
 import * as S from '@/pages/Mypage/MyPage.styled';
-import { UserType } from '@/types/response';
+import { UserType } from '@/shared/types/response';
 
 const FinanceCard = ({ bank, crypto }: Partial<UserType>) => {
   const mask = (num = '') => num.replace(/\d(?=\d{4})/g, '*');
