+++ src/pages/Recipes/index.tsx
@@ -6,7 +6,7 @@ import RecipesItem from '@/pages/Recipes/ui/RecipesItem';
 import { SERVICE_URLS } from '@/shared/api/endpoints';
 import { QUERY_KEYS } from '@/shared/query/key';
 import { useFetchQuery } from '@/shared/query/useFetchQuery';
-import { RecipesType } from '@/types/response/recipe.types';
+import { RecipesType } from '@/shared/types/response';
 
 const Recipes = () => {
   const [selectId, setSelectId] = useState<number | null>();
