+++ src/shared/hooks/useLoginMutation.ts
@@ -2,7 +2,7 @@ import { useQueryClient } from '@tanstack/react-query';
 
 import { tokenStorage } from '@/shared/auth/tokenStorage';
 import { useMutate } from '@/shared/query/useMutates';
-import { AuthLoginRequest, AuthLoginResponse } from '@/types/response/token.types';
+import { AuthLoginRequest, AuthLoginResponse } from '@/shared/types/response';
 
 const useLoginMutation = () => {
   const queryClient = useQueryClient();
