+++ src/shared/hooks/useMeQuery.ts
@@ -2,7 +2,7 @@ import { SERVICE_URLS } from '@/shared/api/endpoints';
 import { tokenStorage } from '@/shared/auth/tokenStorage';
 import { QUERY_KEYS } from '@/shared/query/key';
 import { useFetchQuery } from '@/shared/query/useFetchQuery';
-import { UserType } from '@/types/response';
+import { UserType } from '@/shared/types/response';
 
 const useMeQuery = () => {
   const token = tokenStorage.get();
