+++ src/shared/types/response/products.types.ts
